debos/scripts: Make root during initramfs ro
This really is required for the remount-less rootfs resize, as
resizes on rw mounted filesystems aren't very safe.

We also need to make an empty /etc/machine-id in this situation for
systemd to be able to bind-mount over it from a tmpfs so that
dbus-broker doesn't faceplant in this case.