README: Update TODO list to reflect changes in direction
Instead of shipping our own kernel, so much is mainline that the
only really missing bits are device trees and getting Debian to flip
on some config options. The latter can be done by asking them, for
the former it's best if we have a separate channel through which
we can release DTBs, namely a package in a third party repo.