Add .deb archive generation script
All the debian tools have layers to them, like onions, except all
the layers are made of 20 year old faeces, overcomplicated, and very
geared towards a manual workflow that is not suitable for a simple
CI generated package like we need.

So I rolled my own script to do this. Turns out .debs are just
created by ar with files in a certain order, and within them are
compressed tarballs named control and data with the bits we care
about contained within.

It's 45 lines with whitespace, including the control file template,
which is precisely the sort of simplicity I'm looking for, not yet
another useless wrapper that does the wrong thing for my use-case.