workflow/build_image: Use newer (different) ram init bin
Change from the 1560MHz v1.11 ram init to the 1056MHz v1.13 ram init
binary.

Reportedly, the performance is about the same, but also reportedly,
the power draw is less. I could not reproduce the power draw in my
own tests, and did not bother testing performance, but this is the
right DDR bin to use for RK3566 anyway (the other was an RK3568
targeted binary, which might be relevant).