debos/overlays/resizerootfs: Explicitly copy 'fsck' into the initramfs
While `fsck.ext4` is part of the initramfs, it turns out that `fsck`
itself isn't!
While it normally gets included in the initramfs, for some reason that
doesn't happen with our image creation, so it isn't available on first
boot, so explicitly copy `fsck` so that it will be included.
For that to work, `logsave` is needed too, so include that as well.

During boot up, initramfs wants to check the filesystems and it does so
via `fsck`, which then checks the filesystem being used and calls the
appropriate fsck.* to do the actual checking.
But when `fsck` itself isn't available you get the following warning:
"Warning: fsck not present, so skipping root file system"
When it is available, you'll get a message like this:
"Begin: Will now check root file system ... fsck from util-linux 2.38.1"