debos/overlays: add apt overlay for Experimental
Debian will first release version 6.1~rc5 to Experimental and this
kernel version should have all the needed parts to have a working
Quartz64 device.

So add an overlay for APT which enables the Experimental release.
Also add an APT preferences file which sets the Pin-Priority of
Experimental to '101' (vs the default of '1').
By default, with a value of '1', packages from Experimental will not be
upgraded automatically when a new version is released (to Experimental).
With a value of '101' they will get upgraded.

Closes: #1