debos/overlays/apt: Enable Sid/Unstable release
But also add a configuration file which makes 'Bookworm' the default.
To install something from Sid, do `apt install <pkg> -t sid`.

Without this configuration file, users would be running Sid as Sid has
higher versions of packages and higher versions 'win' when the
(Pin-)Priority is the same, which is (normally) the case with
Sid, Testing or Stable (all '500' by default).