debos: Move repo enablement to top
The files for using the Bookworm security/update and the Sid and
Experimental repos were in place, but APT was aware of them yet.
You'd need to do 'apt update' before it does, but the repos should be
known to APT before the packages are installed, so move the enablement
steps before the "action: apt" step is run.

Also improve some descriptions to be clearer/more correct.
And make explicit that the repos should be updated before the
"action: apt" stage is run.
It is already the default, but explicit > implicit.