debos/overlays/apt: Make '-security' and '-updates' part of Default-Release
https://www.debian.org/releases/bullseye/amd64/release-notes/ch-information.en.html#security-archive
describes how to add bullseye-security and bullseye-updates to the
Default-Release by using APT's support for Regular Expressions.

Without this change, the package version in the normal Bookworm archive
area has a higher APT priority then one in -security or -updates, which
means it won't install (security) updates.
With this change, the highest version wins, which is what we want.