debos: Add initial set of firmware packages
The `bluez-firmware` package is highly likely needed for bluetooth
support, so add that.
The `firmware-brcm80211` package is likely needed for support for some
wifi devices, so add that too.

There is a reasonable chance more firmware packages are needed for
proper/full wifi support, but these 2 are almost certainly needed.