README: Clarify that device tree overlays are still TODO
I thought we'd have these working for sure now, but a cheeky missing
DTC_FLAGS=-@ is throwing a wrench into the works. Since Debian is in
a freeze now, it's too late to get this enabled on their side, but
we can still ship our own device tree package eventually.