workflow: Split device tree package into its own workflow
In preparation for making this into a thing that publishes to a
repository, we split it into a separate workflow.

The reason why it's manually triggered rather than triggered based
on a submodule path receiving a push is that I don't want to have
the huge kernel git as a submodule, and shallow clones with non-
default branches get weird. If you -b foo --depth=1 the linux repo,
it won't be able to find "foo" because the shallow clone doesn't
include this reference. Very annoying, but we'll cope.